[CS-2504] - Update Merchant Balances to look like depot balance
